enhanced app architecture :star:
