+++ app/src/main/res/values/strings.xml
@@ -26,4 +26,7 @@
     <string name="enter_github_username">Enter GitHub username</string>
     <string name="enter_skillsets">Enter Skillsets</string>
     <string name="enter_interests">Enter Interests</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
